add a sweet alert inside the 'Overlay' for when the user clicks on the delete button to confirm their decision

+++ src/components/UI/Overlay.js
@@ -1,4 +1,5 @@
 import React, { useCallback, useEffect } from "react";
+import Swal from "sweetalert2";
 
 const Backdrop = ({ closeModal }) => {
   return <div onClick={closeModal} className="backdrop"></div>;
@@ -12,14 +13,34 @@ const Modal = ({ closeModal, curTask, setTasksHandler }) => {
   const isTimeSpecified = curTask.hours || curTask.minutes || false;
 
   const deleteTaskHandler = () => {
-    const filteredArr = JSON.parse(localStorage.getItem("tasks")).filter(
-      (task) => task.id !== curTask.id
-    );
+    Swal.fire({
+      title: "Estás seguro?",
+      text: "No podrás recuperar la tarea una vez eliminada!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, borrala!",
+      customClass: "swal",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const filteredArr = JSON.parse(localStorage.getItem("tasks")).filter(
+          (task) => task.id !== curTask.id
+        );
 
-    localStorage.setItem("tasks", JSON.stringify(filteredArr));
-    setTasksHandler(filteredArr);
+        localStorage.setItem("tasks", JSON.stringify(filteredArr));
+        setTasksHandler(filteredArr);
 
-    closeModalHandler();
+        closeModalHandler();
+
+        Swal.fire({
+          title: "Borrada!",
+          text: "La tarea ha sido borrada.",
+          icon: "success",
+          customClass: "swal",
+        });
+      }
+    });
   };
 
   return (
